tdaunif: Uniform manifold samplers for topological data analysis

Generate uniform random samples from embedded manifolds, optionally with noise.

Details

This package assembles functions that generate samples of points uniformly from the surfaces of embedded manifolds. An embedding is a one-to-one continuous map \(f:M\to X\) from a manifold \(M\) to a Euclidean coordinate space \(X\), and each function relies on a parameterization of \(M\) given by a continuous bijective function \(p:S\to f(M)\) that may identify some points of \(s\) (boundary or interior) to produce the topology of \(M\). (This means that the inverse of \(p\) may not be continuous.)

Sampling points \(P\) uniformly from \(S\) and mapping the sample to \(f(M)\) may produce a non-uniform sample \(p(P)\) due to differences in the local sampling rate per unit interior (length, area, volume, etc.), quantified as the Jacobian (higher-order derivative) of \(p\). tdaunif uses two techniques to correct for this:

The more numerical (brute-force) technique is to compute the Jacobian on the parameter space and oversample locally at a rate proportional to the Jacobian. This oversampling is done via rejection sampling as illustrated by Diaconis, Holmes, and Shahshahani (2013).

The more analytic technique is to invert the Jacobian symbolically in order to define an interior-preserving parameterization \(q:S\to f(M)\), as illustrated for 2-manifolds by Arvo (2001). Sampling \(P\) uniformly on \(S\) then produces a uniform sample \(q(P)\) on \(f(M)\). The interior-preserving map also enables stratified sampling on the manifold via stratification of the parameter space.

Multivariate Gaussian noise in the coordinate space can be added to any sample.

References

J Arvo (2001) Stratified Sampling of 2-Manifolds. SIGRAPH 2001 (State of the Art in Monte Carlo Ray Tracing for Realistic Image Synthesis), Course Notes, Vol. 29. https://www.cs.princeton.edu/courses/archive/fall04/cos526/papers/course29sig01.pdf

P Diaconis, S Holmes, and M Shahshahani (2013) Sampling from a Manifold. Advances in Modern Statistical Theory and Applications: A Festschrift in honor of Morris L. Eaton, 102--125. doi:10.1214/12-IMSCOLL1006

See also

Author

Jason Cory Brunson

Brandon Demkowicz

Sanmati Choudhary